requirements.txt
aiokafka==0.13.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.31.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
Deprecated==1.3.1
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
idna==3.11
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.25.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pylogbeat==2.1.0
python-dotenv==1.2.2
python-json-logger==4.1.0
python-jose[cryptography]
python-logstash-async==4.1.0
requests==2.33.1
ruff==0.15.10
passlib
SQLAlchemy==2.0.49
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
wrapt==2.1.2
httpx